docs(intranet header): add intranet header showcase (#17)
* docs(intranet header): add intranet header showcase

Migrating the intranet header styles from web-styles to the intranet header angular component had some complications and was therefore abandoned. As the intranet header is due to be replaced by a web component, it's not worth the effort. To still be able to fix bugs, a simple showcase app was added to the stack.

* fix(intranet header): add search option

* build: ncp is no longer required with the new deploy process

* build: move styles to non-peer deps

This prevents pnpm from building the demo and the header in parallel during bootstrap.