feat(colors)!: refactored internal color usage (#96)
This changes separates post and bootstrap color names. Post colors are used throughout the CWF while the bootstrap color file only overrides bootstrap colors and is only used to compile bootstrap with post colors. This leads to a clear separation of colors and disables access to default bootstrap colors from within the CWF.