chore(form-select/custom-select): Fixes floating-label visibility, when rendered without option-tags or without selected option (#169)
The native form-select element shows the text of the first option-tag, if no option is selected. The floating-label comes only in place (like a placeholder) if there is no option-tag present.
In contrast, the custom-select element shows the floating-label like a placeholder, if there is no option-tag available, or none of its options is selected.