fix(alert): icons in high contrast mode (#351)
Alerts with custom icons face the problem that those icons are black. Inversion techniques don't work because it would invert the whole alert. Creating white alternatives for all possible icons would blow up CSS size too much with current approaches so it's not viable. This approach just replaces custom icons with the default alert icon. This way, only a small set of white icons has to be created.

The linter seems broken when dealing with placeholders, but it's an unusual pattern anyways. Created a ticket (#352) to remedy this situation.