feat(naming)!: apply new naming conventions (#355)
Following the discussion at #304, some things need to be renamed for maximum consistency.

This change was mainly triggered in an effort to align development resources with the design and the documentation. Design System is the officially coined term and thus a perfect way to describe what it is by simply naming it. Swiss Post gives the package the correct scope.

Other namings were also aligned with this descision, for example package names:

Pattern: [company scope]/[project]-[target platform]-[package name]-[modifier/s] (the web platform is implied as default platform).

@swisspost/design-system-components
@swisspost/design-system-components-angular
@swisspost/design-system-components-vue
@swisspost/design-system-header
@swisspost/design-system-intranet-header
@swisspost/design-system-migrations
@swisspost/design-system-styles

The short form post was chosen because of its superior recognizability over any abbreviation. This will be used to prefix CSS variables (#358) and web components.