fix: add build script to intranet-header package (#655)
The components angular project repo is the place where all the node modules for the angular components are installed. When filtering with pnpm dependency chains (e.g. in the test workflow -> ...design-system-demo...), the components-angular project is not part of the dependency chain and therefore the node modules are not being installed and the header fails to compile.

By adding the components angular project as a dev dependency to the intranet header component, component-angular becomes part of the dependency chain, node modules get installed and the build script is being executed. Problem solved? We'll see.

Co-authored-by: Philipp Gfeller <philipp.gfeller@post.ch>