build(release): check if package version updated for docs deploys (#711)
The old check only checked if any package got released to npm, which is not the case if the only the documentation got updated. This did not trigger a deploy of the documentations. The intended behaviour was to check if the demo package.json version has changed and do a release then.