feat(styles-migration): create migrations with the new cheerio util (#687)
This PR adds new schematic utilities, to write easy jQuery-like migrations for almost all necessary DOM migrations.
And replaces all former existing with new migrations based on the new utilities.

Under the hood we use angular schematics built-in template visitor and cheerio to access the html in a jQuery-like context. Updates are formatted with prettier and commited directly on the tree files.

We added an example migration to show how to write new migrations.

To learn how to execute the schematic migrations, visit https://design-system.post.ch.