refactor: strict null checks and linting errors
Enabled strict null checks and refactored the code accordingly. Also fixed a few linting errors along the way. Some stencil rules are buggy with the eslint 8 compiler and some react/jsx-no-bind warnings are still to be fixed but are a bit more complex because they pass custom parameters to their handlers that are not available outside the render function. Maybe the compnents could be re-structured to no longer depend on these values, but that's something for another day.