fix(internet-header): accessibility issues and broken config (#1066)
* fix(header): hidden button text was inaccessible

Visibility: hidden prevents screen readers from reading the text content of an element. Removed the style, now the standard bootstrap visually hidden styles are applied.

* fix(header): update skiplinks markup

Updated skiplinks markup to better match the previous implementation and prevent an unnamed landmark error. Also added a title.

* fix(header): header in custom config is optional

Fixed an issue with the footer custom config. The header part of the custom config is optional and is now treated as such.