refactor: scripts & tests (#1097)
* refactor(header): move cypress test to package

De-centralizing cypress tests has the benefit of being able to run tests only for the packages that changed instead of everytime for all the package changes. Cypress tests are quite heavy and can run for a long time.

* refactor: script naming and availability at root

This largely refactors and documents how scripts are named and implements a lot more scripts available at the root of the project.

* chore: remove chromatic deploy

For the moment, there's no need to publish to chromatic. As snapshots are not very useful right now, there is no reason to further investigate the other possibilities of the platform.

* tests: add button snapshot page and test

The button snapshot page is a single story listing all possible combination of buttons on one page. Pretty neat!