feat(stepper)!: Updated stepper markup (#1232)
* feat(stepper)!: Updated stepper markup

Updated the markup for the stepper component in order to improve accessibility. Completed items now require the prefix "Complete" and incomplete steps are no longer links.

* fix: progress bar color

The progress bar can be colored using bootstrap backgrounds. When changing the info color to light blue, there is no longer any bootstrap background with the brand yellow. Therefore we have to override the active progress color with yellow.