fix(header): improve tag name choice heuristic
Links without href are not tabbable by default. If there is no url or an empty one, a button should be rendered for the language switch. The header will then use the language change event for communicating with the page about language changes (this should be the exception in an online service as most of them have language URLs defined).