test: update stickyness test
After scrolling, wait for elements that should be in the viewoprt instead of just checking if some elements are not in the viewport. When just checking the latter, cypress runs the check during scroll and it's immediately successful so the next scroll is executed. If there is an element that should be there, cypress has to wait for the scroll to finish before the element appears.