fix(header): track and trace redirect logic
The track and trace API changed it's ok response type from boolean to string which could not be handled by the redirect logic of the header. Requests that matched the very generic regex were redirected to track and trace, even if the search term was not a tracing number. The new logic checks if there are results instead of relying on the ok property.

Also, the string format of the parcel info preview was updated to account for missing information.