fix(styles): replace jest by scss alone and make tests work again (#1662)
Currently, the command jest --OnlyChanged using jest-scss-transformer.js is not working. The newest version of sass requires node packages like fs to read the input file, but it doesn't have access through jest. Discussed with @gfellerph, here jest is an unnecessary overhead. The only benefit is the colored syntax in the terminal for errors.