chore: improve styles watch performance
Only copies changed files to dist. This will improve performance for development massively because vite reacted to every change in dist individually and before, every single sass file was copied to dist regardless if it changed or not. Also, map-icons creates a theoretical infinite loop and calling watch again is not necessary within the watch task.