feat(components): tooltips (#1879)
Adds a `post-tooltip` component.

- Why `[popover]`?: using the popover attribute (with polyfill) has
several advantages. In supporting browsers, the tooltip will be lifted
to the top-layer, ignoring `overflow: hidden` and the `z-index` order.
This ensures that tooltips are always displayed on top of everything and
never cut off. Until all browsers support the popover API (ETA is end of
2023), a polyfill without these qualities (but some decent standards) is
needed.
- Why `[data-tooltip-trigger]`? This is a custom attribute in place of
`popovertargetaction`, an attribute that can be used to trigger a
popover, but reacts to clicks only. In this situation we want to show
the popover on hover, focus and long-press. This behavior is in
consideration with `popovertargetaction="interest"`, but is not
implemented at the moment. To be able to patch this functionality
without interference from clicks, there is no `popovertargetaction` on
the trigger element.
- floating-ui: this component uses
[floating-ui](https://floating-ui.com/docs/getting-started), the
successor to `@popperjs/core`, for positioning. This low-level library
can be used to dynamically position the tooltip and the corresponding
arrow and can be customised in every imaginable way while leaving a
pretty small footprint.
- No snapshot tests: not sure how to test something that only shows up
on user interaction
- No animation: had some trouble implementing an animation that was
compatible with the polyfill as well as the default API. Suggestion:
wait till Design provides a concept and the popover API is standardised

---------

Co-authored-by: imagoiq <12294151+imagoiq@users.noreply.github.com>