fix(documentation): Remove incompatible vite version (#2083)
https://github.com/swisspost/design-system/pull/2063 introduces another
`vite` package version (v4.4.11) which is not the same as the one behind
Storybook (v4.4.7). It seems that there are incompatibilities as the
`*.entry.js` files of the components packages are not built with this
change.

This PR removes this `vite` package and the code relies on the
sub-dependency `vite` package of storybook.