fix(e2e): Revert vite 4.11 upgrade (#2160)
https://github.com/swisspost/design-system/pull/2072 introduces some
sub-dependencies changes similarly to the initial PR of this revert:
https://github.com/swisspost/design-system/pull/2083. This breaks e2e
tests for components and internet-header. `*.entry.js` files are not
generated/copied to the documentation assets folder and creates a 404
when we try to use a web component.

For testing, I cherry-pick the commit of this PR to
https://github.com/swisspost/design-system/pull/2140

Also upgrading to Stencil 4
(https://github.com/swisspost/design-system/pull/2116) does not fix this
issue.