feat(documentation): Add placeholder feature for floating-label select (#2200)
Add a way to see label as a placeholder for select element when
`floating-label` is enabled.

Currently, we can use the `select:empty` selector, remove all options
(so all the children of the select element) and repopulate them when the
user click on the select. But it's quite cumbersome for the end
developer. It's much easier to add a class. In the future, we'll be able
to do that using `:has` instead of class.

Linked to this old issue:
https://github.com/swisspost/design-system/issues/86

This issue fixes as well the color to match the input one.

---------

Co-authored-by: Alizé Debray <33580481+alizedebray@users.noreply.github.com>