chore: zip build artifacts (#2258)
This change will centrally handle zipping and unzipping of build
artifacts for faster upload/download. There is one catch to this PR: the
download action always runs on the main branch (for security reasons)
and cannot be tested, therefore the documentation publishing workflow
fails for this PR. To actually test it, we'd need to merge this PR and
push a change to the documentation.