fix(styles): Intranet-header sticky nav prevents interactivity (#2320)
This was tested on a side project. The issue is not visible in
https://design-system.post.ch/#/post-samples/intranet-layout because the
components styles have been rewritten to be used inline (position fixed
was removed). It would be nice to rewrite this demo by using iframe
instead, but it's too much work for the deprecated old design system
doc.