fix(components): stencil unit test require puppeteer (#2321)
We had the same issue some weeks ago. e.g. in this PR:
https://github.com/swisspost/design-system/actions/runs/6785015221/job/18442967411?pr=2065#step:5:19
And we fixed with:
https://github.com/swisspost/design-system/pull/2238/files#diff-4ade8993db0ab98cef66268e7a71355b158eefa4db4a3e03f623638800612980

But it came back, and it seems it's linked to
`createJestPuppeteerEnvironment`. I've found a PR and an issue on
stencil/ionic side which seems to indicate that we need Puppeteer:
https://github.com/ionic-team/ionic-framework/pull/28529#discussion_r1394451290
https://github.com/ionic-team/stencil/issues/4526

Tested successfully in this PR:
https://github.com/swisspost/design-system/actions/runs/7043980625/job/19170759773?pr=2320