fix(styles): button group do not have focus style for checkbox/radio variants (#2273)
The solution works for the checkbox. For the radio variant, since with
tab you can only focus the first element of the radio button group, it
also only shows the focus style of the first label. If you want to focus
the others, you have to use the arrow keys. And if I would put all the
radio buttons in a seperate group, we would loose the whole
functionality of the component. At least for chrome its like that, in
firefox it works.

---------

Co-authored-by: Loïc Fürhoff <12294151+imagoiq@users.noreply.github.com>