feat(docs): Add autolink headings (#2467)
This is proof-of-concept. It has some complexities, and please have no
hesitation to say if it's preferable not to add it.

HTML markup for links is added through
[remark-autolink-headings](https://github.com/remarkjs/remark-autolink-headings).
Remark plugins are officially supported by Storybook Docs and can be
easily added.

### First issue (legacy plugin)

However, this plugin is considered as legacy and not updated since 2021.
The author recommends using `remark-rehype` and
`rehype-autolink-headings` together instead. Unfortunately, it's not
possible with our setup as it'll create a double transformation into
HTML as Storybook is already doing a transformation. I think that it's
not a showstopper, as the plugin is quite simple (1 file) and could be
even integrated into our repository if needed.

### Second issue (monkey patch)

Another requirement was to add a script that rewrite the `base` element
of the preview iframe to be able to get the anchor link related to the
main URL and not the iframe. Without it, it would work, but let the user
without sidebar and toolbar, and it would be just confusing for the
user.

I think, it's quite unlikely that Storybook will update themselves the
`base` element or let us do it as [the element is included in a basic
HTML
file.](https://github.com/storybookjs/storybook/blob/3664135b74c589f5c24bbfbad4438fd3fe32cd74/code/lib/core-common/templates/base-preview-head.html#L4)

---------

Co-authored-by: Oliver Schürch <oliver.schuerch@post.ch>