fix(documentation): Autolink with base url disrupt other anchor links (#2529)
Fix over https://github.com/swisspost/design-system/pull/2467

In the previous PR, autolink anchor link associated to heading were
formed by changing the `<base>` element href attribute so that the
absolute URL would be created by using the parent frame URL (the main
URL that the user navigate with). However, it appears that other anchor
links, mostly the one used for the table of content, will use this
absolute URL as well. They are correct, but the transition to an anchor
is managed by Storybook and not the browser so in this case as the URL
used is not the same as the iframe, it doesn't work if the page is
already loaded (it works when the page is refreshed).

To avoid this issue, I propose a progressive enhancement to capture all
the autolink anchor links and change them (but only them) to use the
absolute URL from the main frame. If the script fails, the link still
works but the manager (sidebar, toolbar) of Storybook will not appear.

Ideally, we should generate those links in Storybook. But I couldn't
find a way to get the context in the rehype plugin and hooking into the
layout.tsx children don't seem feasible.

### Testing

You can find table of contents in
https://preview-2529--swisspost-design-system-next.netlify.app/?path=/docs/components-icons-getting-started--docs
(although some links are broken, but fixed with
https://github.com/swisspost/design-system/pull/2496)