feat(setup): update netifly to v16 (#2421)
![image](https://github.com/swisspost/design-system/assets/141234680/4b539781-c3f7-441f-bf09-492d855ef724)

I think if we update our deploy-to-netifly action to use a filter with a
package name we pass to this action then using netifly v16 would be no
Problem.
However @oliverschuerch and myself @b1aserlu failed to come up with a
way to properly test this. As for some reason the deploy worklfows are
only triggered from main and not as expected from the current branch.