fix(tooltip): don't move around the tooltip
Moving the tooltip out of the current context means frameworks like Angular are no longer able to track it. This leads to strange errors where Angular router navigations re-instantiate the tooltip and it gets added to the <body> directly. Upon multiple navigations, more and more tooltips are being added to the end of the body and Angular can't remove the unused ones.

Overflow hidden issues can be solved by manually placing the tooltip outside of the overflow container.