fix(popovercontainer): prevent immediate re-toggle (#2695)
This small timeout prevents the popovercontainer from opening and
closing with the same event chain. This happens for example when the
popover is open and the trigger button is clicked again. It's because
first, an outside click is registered the popover closes. Right after
that, the click from the toggle button fires and opens the popover
again.