fix(styles): Color-contrast on post-tabs with dark background (#2753)
1. Fixed color-contrast of post-tabs on dark-background:

From

![Cypress_3Nn8pl3IFf](https://github.com/swisspost/design-system/assets/12294151/6098f7ed-7be5-41c6-ae88-33af32f8cc76)

To 
![Screenshot 2024-03-05 at 08-33-23
snapshots--tabs](https://github.com/swisspost/design-system/assets/12294151/118e3ffe-397c-4875-b569-b96d388e294d)


2. Added cursor pointer to tabs title
3. Revert change from
https://github.com/swisspost/design-system/pull/2349/files/344934ab280122c643d2886666f3981829ff2ff1#r1413888274
and fix with added spacing. The margin is visible here:


![image](https://github.com/swisspost/design-system/assets/12294151/08e89b45-f1e6-4b80-bf64-c4dbd2ef83f7)

Note: @rouvenpost @Cian77 I didn't implement what's in the design
https://www.figma.com/file/xZ0IW0MJO0vnFicmrHiKaY/Components-Post?type=design&node-id=70-18&mode=design&t=VxSxr2DejdjVc4mY-0
because it misses how it should look like on dark colored background and
because we are quite limited at the moment with this `post-tabs`
component as the `:host-context` selector is not available (and will
never be it seems) on all browsers, so to adapt to a particular theme
given by the parent, we can only use CSS custom properties (like
`--post-contrast-color-inverted` or `--post-contrast-color`). That means
the component doesn't know anything about the current theme. This is
good if we want to go further on theme architecture where you only need
to provide different variables to change a component, but we are not
there yet. And finally, it is also complex to style for the `post-tabs`
component as well for the `tabs` HTML component, perhaps we should
decouple those two.