fix: define CSS custom properties for header/crumbs/footer (#2793)
If the Internet Header was used without having the styles as a global
stylesheet, some CSS custom properties were not defined. This lead to
focus rings not being visible. Extending the light theme colors for all
the entry components defines these properties even if the styles package
is not being used.