chore(versioning): fix versions of main packages (#2856)
This will start to align the version of the fixed packages together.
Fixed packages will be published under a new version, even if they did
not have any changes
(https://github.com/changesets/changesets/blob/main/docs/fixed-packages.md).

There are several packages, that are not included in the list because
they should still be versioned individually. The reason for this is,
that they are not main deliverables of the Design System - styles and
components. For example the documnetation, which can only be released
with a version bump. If it would be fixed with the other packages, a
correction of a typo would cause a version bump for all the
deliverables, even if nothing changed for end users.