fix(styles): Alert actions button color contrast was not met with success background color (#2845)
### What

From design files, text color of alert actions should be white on dark
background (currently, it's 80% white as it's based on the
`--post-black-80` variable, but inverted). To define another color, I
needed to add the alert classes and selector (for the web component)
into the dark-backgrounds variables. The notification name list was
moved to color file to avoid recursion issue (perhaps we could put it
somewhere else).

I'll check with design what to do with hover style.