chore: remove docs peer dependencies
A weird workflow (https://github.com/changesets/changesets/issues/1011) in changesets causes packages depending on others in the monorepo through peer dependencies to cause a major update on the upstream package. This is undesired in this scenario. Example:

design-system-components-angular has a patch update and is a peer dependency of the documentation. Because of that patch, the documentation has a major version bump for this release, even if there are no breaking changes.