chore: fix pnpm start command (#3504)
This change excludes `@swisspost/design-system-styles` from the vite
optmizeDeps workflow for the documentation package. This seems to get
rid of some of the "can't import dynamic dependency" issues.

Reloading after a style change is still pretty slow, but also a lot of
unnecessry recompiles are happening. The potential for improvement is
quite big here.