chore: set a fixed netlify cli version (#3593)
Our last successful deploys were on the 20th of September. On that day,
netlify released a patch update that seems to have broken the deploy
command if netlify builds are disabled
(https://github.com/netlify/cli/blob/main/CHANGELOG.md#17362-2024-09-20).
Trying to fix the issue by setting netlify version to a version before
that.

Submited an issue for netlify cli:
https://github.com/netlify/cli/issues/6841