fix: release workflow publish with provenance (#4580)
Regarding the
[docs](https://docs.npmjs.com/generating-provenance-statements#publishing-packages-with-provenance-via-github-actions)
this should be enough to publish packages with provenance.
But it could also be a configuration problem in the
SWISSPOSTDEVS_ACCESS_TOKEN, which is used to call the changesets/action
in our release workflow.

@gfellerph Let's try this first, before we switch to the GITHUB_TOKEN.
In case this will not work as well, I'll remove the provenance settings
for the moment, so the release is not blocked anymore.