chore(components): fixed popover not opening when clicking on elements nested in trigger (v8) (#4911)
## 📄 Description

Changed the globalToggleHandler method of the post-popover component.
Before it only checked if the event target had the 'data-popover-target'
attribute. When for example an icon is nested in the trigger, this would
result to false and the popover would not open. In the new
implementation instead of just checking the e.target we traverse the DOM
tree upwards searching for an element with the 'data-popover-target'
attribute. Now clicking on an icon nested in the popover trigger, the
popover will be opened/closed as expected.

## 🚀 Demo



https://github.com/user-attachments/assets/6519dcde-f1d2-4354-a324-6f9c2609f7bd


---

## 📝 Checklist

- ✅ My code follows the style guidelines of this project
- 🛠️ I have performed a self-review of my own code
- 📄 I have made corresponding changes to the documentation
- ⚠️ My changes generate no new warnings or errors
- 🧪 I have added tests that prove my fix is effective or that my feature
works
- ✔️ New and existing unit tests pass locally with my changes