chore(components): expose header height to root (#4925)
## 📄 Description

This PR moves the `--post-header-height` (visible height) and
`--post-header-expanded-height` custom properties to the `:root` via the
header/index.scss. All the custom properties involved and thus also
exposed to the `:root` have also been updated with the `post-` prefix.
Additionaly, the `scroll-padding-top` property is set as the value of
the (visible) header height.

## 📝 Checklist

- ✅ My code follows the style guidelines of this project
- 🛠️ I have performed a self-review of my own code
- 📄 I have made corresponding changes to the documentation
- ⚠️ My changes generate no new warnings or errors
- ✔️ Existing unit tests pass locally with my changes

---------

Co-authored-by: Oliver Schürch <oliver.schuerch@post.ch>