chore(tests): added header tests simulating react navigation behaviour (#5098)
## 📄 Description
Added two tests simulating react navigation behavior by removing and
reattaching the header from the DOM. The first test checks that the
mobile menu is correctly closed after the header is reattached and the
second one checks that the scroll lock is added and released at the
correct times.

There was no way to test the correct removal of event listeners. I tried
all the following approaches:

- Use the getEventListener available in chrome dev environment
- Spy on function calls to Add & RemoveEventListener
- Overwrite Add & RemoveEventListener with trackable version
- Check heap space usage to detect memory leaks

## 🚀 Demo

Not really helpful here.
---

## 📝 Checklist

- ✅ My code follows the style guidelines of this project
- 🛠️ I have performed a self-review of my own code
- 📄 I have made corresponding changes to the documentation
- ⚠️ My changes generate no new warnings or errors
- 🧪 I have added tests that prove my fix is effective or that my feature
works
- ✔️ New and existing unit tests pass locally with my changes