chore(components): fix logo disappearing on page locale change (#5268)
## 📄 Issue Description

The logo disappearing issue is caused by the missing custom property
`--post-header-scroll-top` within the `--logo-height` calculation. The
property has been so far set in the connectedCallback.

## Testing
When tested within the Cargo implementation:

a) on first load it works as expected
b) on language /country change the page is re-rendered but the
post-header connectedCallback seems to be called before the React
renders (tested with useEffect() inside the PostHeader component) and
the custom property is unset.
c) When moved to componentWillRender the custom property is set
successfully.

## Proposed Solution

This PR moves the handleScrollEvent function from the connectedCallback
to the componentWillRender method. Additionally, a fallback value is
added for the `--post-header-scroll-top` within the `--logo-height`
calculation.

## 📝 Checklist

- ✅ My code follows the style guidelines of this project
- 🛠️ I have performed a self-review of my own code
- ⚠️ My changes generate no new warnings or errors
- ✔️ New and existing unit tests pass locally with my changes