chore(documentation): remove ng-bootstrap custom select (#5302)
## 📄 Description

Remove documentation about the ng-bootstrap custom select.

Note that the changeset is set to `minor` as it's only the documentation
that has been updated. Technically, all the styles are still present as
they are common with the `form-select`, and the JS comes from importing
ng-bootstrap which we have not removed.