feat(documentation): deep link to specific component configuration (#4828)
This PR goal is to add to the storybook 'docs' mode the ability to link
from a component page to another one with specific args set. (e.g.
button-secondary). With the current PR changes the args are applied to
the component via the use of a global decorator which takes the URL
params and updates the args. (e.g. this link will show the secondary
button:
http://localhost:9000/?path=/docs/eb78afcb-ce92-4990-94b6-6536d5ec6af4--docs&args=variant:btn-secondary).

A 'Copy variation' button is also added, to copy the current
configuration link to the clipboard.

---------

Co-authored-by: Philipp Gfeller <1659006+gfellerph@users.noreply.github.com>