chore(docs): move focus to icon details popover to ensure accessible keyboard navigation (#5365)
## 📄 Description
When opening the icon details, the focus now moves to the first
focusable element within the popover. Currently this is the download
anchor element. When pressing tab again it moves to the next focusable
element which is the close button. After that it moves back to the first
essentially trapping the focus inside the popover until it is closed and
the focus is returned back to the element which opened the details.

## 🚀 Demo

Uploading FindIconFocusBehaviour.mp4…

---

## 📝 Checklist

- ✅ My code follows the style guidelines of this project
- 🛠️ I have performed a self-review of my own code
- 📄 I have made corresponding changes to the documentation
- ⚠️ My changes generate no new warnings or errors
- 🧪 I have added tests that prove my fix is effective or that my feature
works
- ✔️ New and existing unit tests pass locally with my changes