chore(setup): upgraded `styles:build` task to use turborepo (#5341)
## 📄 Description

This PR is the result of the POC for Turborepo. It adds turborepo as a
dependency and upgrades the `styles:build` command to make us of local
computation caching. To enable the caching the build task was defined in
the turbo.json file with all the outputs which need to be cached for the
specific package.

## 🚀 Demo
To test local caching, run pnpm styles:build twice in a row :)

---

## 📝 Checklist

- ✅ My code follows the style guidelines of this project
- 🛠️ I have performed a self-review of my own code
- 📄 I have made corresponding changes to the documentation
- ⚠️ My changes generate no new warnings or errors
- 🧪 I have added tests that prove my fix is effective or that my feature
works
- ✔️ New and existing unit tests pass locally with my changes